requirements.txt
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.40.0
gradio>=4.36.0
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
huggingface-hub>=0.23.0
accelerate>=0.30.0
sentencepiece>=0.2.0
scikit-learn>=1.4.0
pytest>=8.0.0
sentence-transformers>=3.0.0
flask>=3.0.0
werkzeug>=3.0.0
